+++ frontEnd/src/config/api.config.js
@@ -1,8 +1,11 @@
-export const API_URL = import.meta.env.VITE_API_URL || 'http://localhost:8000';
+export const API_URL = import.meta.env.VITE_API_URL || 'http://127.0.0.1:8000/api/v1';
 
-// You can also add other API-related configurations here
 export const API_CONFIG = {
     headers: {
-        'Content-Type': 'application/json'
-    }
+        'Content-Type': 'application/json',
+        'Accept': 'application/json'
+    },
+    getAuthHeader: () => ({
+        'Authorization': `Bearer ${localStorage.getItem('token')}`
+    })
 };
