+++ frontEnd/src/services/auth.services.js
@@ -3,13 +3,17 @@ import { API_URL } from '../config/api.config';
 export const authService = {
   async login(username, password) {
     try {
-      const response = await fetch(
-        `${API_URL}/users/login?username=${encodeURIComponent(username)}&password=${encodeURIComponent(password)}`,
-        {
-          method: 'POST',
-          headers: { 'accept': 'application/json' }
-        }
-      );
+      const response = await fetch(`${API_URL}/users/login`, {
+        method: 'POST',
+        headers: { 
+          'Content-Type': 'application/json',
+          'Accept': 'application/json'
+        },
+        body: JSON.stringify({
+          login: username,
+          password: password
+        })
+      });
 
       if (!response.ok) {
         const error = await response.json();
@@ -24,7 +28,7 @@ export const authService = {
 
   async register(userData) {
     try {
-      const response = await fetch(`${API_URL}/users/`, {
+      const response = await fetch(`${API_URL}/users/register`, {
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify(userData)
@@ -43,13 +47,16 @@ export const authService = {
 
   async logout(username) {
     try {
+      // Get the token from localStorage or your auth state
+      const token = localStorage.getItem('token') || '';
+      
       const response = await fetch(`${API_URL}/users/logout`, {
         method: 'POST',
         headers: { 
           'Content-Type': 'application/json',
-          'Accept': 'application/json'
-        },
-        body: JSON.stringify({ username: username })
+          'Accept': 'application/json',
+          'Authorization': `Bearer ${token}`
+        }
       });
 
       if (!response.ok) {
